+++ CHANGES.rst
@@ -4,6 +4,7 @@ Release 9.1.1 (in development)
 Bugs fixed
 ----------
 
+* #14587: Combine format-wide and builder-specific node translation handlers.
 * #14465: LaTeX: PDF build crash since LaTeX June 2026 release if tables are
   styled with ``'colorrows'`` (which is the default).
   Patch by Jean-François B.